lib multi bump var to size_t

+++ lib/multiif.h
@@ -46,7 +46,7 @@ void Curl_multi_connchanged(struct Curl_multi *multi);
 /* Internal version of curl_multi_init() accepts size parameters for the
    socket, connection and dns hashes */
 struct Curl_multi *Curl_multi_handle(int hashsize, int chashsize,
-                                     int dnssize);
+                                     size_t dnssize);
 
 /* the write bits start at bit 16 for the *getsock() bitmap */
 #define GETSOCK_WRITEBITSTART 16
